@@ -7,21 +7,23 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@oruga-ui/oruga": "^0.5.5",
+    "@oruga-ui/oruga-next": "^0.5.5",
     "@oruga-ui/theme-bulma": "^0.2.5",
-    "apexcharts": "^3.35.0",
+    "@soerenmartius/vue3-clipboard": "^0.1.2",
+    "@vitejs/plugin-vue": "^2.3.3",
+    "@vue-leaflet/vue-leaflet": "^0.6.1",
+    "apexcharts": "^3.35.3",
     "axios": "^0.26.1",
     "bulma-block-list": "^0.6.0",
     "dayjs": "^1.11.0",
     "leaflet": "^1.7.1",
     "marked": "^4.0.12",
+    "mitt": "^3.0.0",
     "pinia": "^2.0.14",
     "sortablejs": "^1.15.0",
-    "vue": "2.7.0-alpha.4",
-    "vue-apexcharts": "^1.6.2",
-    "vue-clipboard2": "^0.3.3",
-    "vue-router": "^3.5.2",
-    "vue2-leaflet": "^2.7.1",
+    "vue": "3.2.37",
+    "vue-router": "4",
+    "vue3-apexcharts": "^1.4.1",
     "vuedraggable": "^2.24.3"
   },
   "resolutions": {
@@ -34,9 +36,7 @@
     "sass": "^1.52.1",
     "sass-loader": "^10.2.1",
     "typescript": "^4.6.3",
-    "vite": "^2.9.9",
-    "vite-plugin-vue2": "^1.9.3",
-    "vue-template-compiler": "^2.7.0-alpha.4"
+    "vite": "^2.9.9"
   },
   "browserslist": [
     "> 1%",
