@@ -1,52 +1,36 @@
-import Vue from 'vue'
-import { createPinia, PiniaVuePlugin } from 'pinia'
-Vue.use(PiniaVuePlugin);
-const pinia = createPinia();
-
-// @ts-ignore
-pinia.install(PiniaVuePlugin); // Hack para crear instancia de pinia en vue2 y que el router en beforeEach lo pueda usar
-
-// @ts-ignore
-import VueHeadful from 'vue-headful';
-import VueClipboard from 'vue-clipboard2'
-import Loading from '@/components/lib/Loading.vue'
-
-import './scss/app.scss';
-import { loadFilters } from "@/plugins/dayjs";
-import { loadLeaflet } from "@/plugins/leaflet";
-import mixin from "@/plugins/mixin";
-
-import Oruga from '@oruga-ui/oruga'
+import {createApp} from 'vue'
+import { createPinia } from 'pinia'
+import Oruga from '@oruga-ui/oruga-next';
 import { bulmaConfig } from '@oruga-ui/theme-bulma'
 
-loadFilters(Vue);
-loadLeaflet();
+import './scss/app.scss';
+import {VueClipboard} from "@soerenmartius/vue3-clipboard";
+import VueApexCharts from "vue3-apexcharts";
+import mixin from "@/plugins/mixin";
+import App from "@/App.vue";
+import Loading from '@/components/lib/Loading.vue'
+import router from "@/_router";
 
-import router from '@/_router'
-import App from '@/App.vue'
-
-Vue.use(Oruga, {
-    customIconPacks: {
-        fas: {
-            sizes: {
-                default: '',
-                small: 'fa-sm',
-                medium: 'fa-lg',
-                large: 'fa-2x'
+const app = createApp(App)
+    .use(createPinia())
+    .use(router)
+    .use(Oruga, {
+        customIconPacks: {
+            fas: {
+                sizes: {
+                    default: '',
+                    small: 'fa-sm',
+                    medium: 'fa-lg',
+                    large: 'fa-2x'
+                }
             }
-        }
-    },
-    iconPack: 'fas',
-    ...bulmaConfig
-})
-Vue.use(VueClipboard);
-Vue.config.productionTip = false;
+        },
+        iconPack: 'fas',
+        ...bulmaConfig}
+    )
+    .use(VueApexCharts)
+    .use(VueClipboard)
+    .component("Loading", Loading)
+    .mixin(mixin);
 
-Vue.component("Loading", Loading);
-Vue.mixin(mixin);
-
-new Vue({
-    pinia,
-    router,
-    render: h => h(App)
-}).$mount('#app');
+app.mount('#app');
