@@ -1,8 +1,8 @@
-import {NavigationGuardNext, Route} from "vue-router";
-import {useAuthStore} from "@/pinia/authStore";
+import {NavigationGuardNext, RouteLocationNormalized} from "vue-router";
+import {useAuthStore} from "@/store/authStore";
 
 // Requerir login cuando sea necesario
-export default function checkLoggedIn(to: Route, from: Route, next: NavigationGuardNext) {
+export default function checkLoggedIn(to: RouteLocationNormalized, from: RouteLocationNormalized, next: NavigationGuardNext) {
     const authStore = useAuthStore();
 
     if (to.matched.some(record => record.meta.requiresAuth)) {
