@@ -333,7 +333,7 @@ export const routes = [
         }]
     },
     {
-        path: '*',
+        path: '/:pathMatch(.*)',
         redirect: { path: "/404" }
     }
 ];
