@@ -24,8 +24,8 @@ function authHeader() {
 
 axios.defaults.baseURL = import.meta.env.NODE_ENV === 'production'
     ? 'https://api.prognosports.com/v2'
-   // : 'https://api.prognosports.com/v2';
-     : 'http://localhost:8001/v2';
+    : 'https://api.prognosports.com/v2';
+   //  : 'http://localhost:8001/v2';
 
 axios.interceptors.request.use(function (config) {
     const token = authHeader();
