@@ -59,7 +59,7 @@ import {constructorService} from "@/_services";
 import {Constructor} from "@/types/Constructor";
 
 import {defineComponent} from "vue";
-import {useAuthStore} from "@/pinia/authStore";
+import {useAuthStore} from "@/store/authStore";
 
 export default defineComponent({
     name: "LandingNavbar",
