@@ -72,7 +72,7 @@ import {ICompetition} from "@/types/Competition";
 import {marked} from "marked";
 
 import {defineComponent} from "vue";
-import {useAuthStore} from "@/pinia/authStore";
+import {useAuthStore} from "@/store/authStore";
 
 export default defineComponent({
     name: "SeasonCreate",
