@@ -76,7 +76,7 @@ import {Season} from "@/types/Season";
 import {marked} from "marked";
 
 import {defineComponent} from "vue";
-import {useAuthStore} from "@/pinia/authStore";
+import {useAuthStore} from "@/store/authStore";
 
 export default defineComponent({
     name: "CompetitionEdit",
