@@ -75,7 +75,7 @@ import {competitionService} from "@/_services";
 import {Competition} from "@/types/Competition";
 
 import {defineComponent} from "vue";
-import {useAuthStore} from "@/pinia/authStore";
+import {useAuthStore} from "@/store/authStore";
 
 export default defineComponent({
     name: "CompetitionsList",
