@@ -92,7 +92,7 @@ import {GrandPrix} from "@/types/GrandPrix";
 import {Circuit} from "@/types/Circuit";
 
 import {defineComponent} from "vue";
-import {useAuthStore} from "@/pinia/authStore";
+import {useAuthStore} from "@/store/authStore";
 
 export default defineComponent({
     name: "CompetitionEdit",
