@@ -36,8 +36,8 @@
     import {hasVariant} from "@/utils";
 
     import {defineComponent, PropType} from "vue";
-    import {useAuthStore} from "@/pinia/authStore";
-    import {useCommunityStore} from "@/pinia/communityStore";
+    import {useAuthStore} from "@/store/authStore";
+    import {useCommunityStore} from "@/store/communityStore";
 
     export default defineComponent({
         name: "CircuitCard",
