@@ -22,8 +22,8 @@
     import {grandPrixService} from "@/_services";
 
     import {defineComponent, PropType} from "vue";
-    import {useAuthStore} from "@/pinia/authStore";
-    import {useCommunityStore} from "@/pinia/communityStore";
+    import {useAuthStore} from "@/store/authStore";
+    import {useCommunityStore} from "@/store/communityStore";
 
     export default defineComponent({
         name: "GrandPrixesList",
