@@ -25,8 +25,8 @@ import {StartGridPosition} from "@/types/StartGridPosition";
 import {Driver} from "@/types/Driver";
 
 import {defineComponent, PropType} from "vue";
-import {useAuthStore} from "@/pinia/authStore";
-import {useCommunityStore} from "@/pinia/communityStore";
+import {useAuthStore} from "@/store/authStore";
+import {useCommunityStore} from "@/store/communityStore";
 
 export default defineComponent({
     name: "StartGridCard",
