@@ -37,7 +37,7 @@
     import {seasonService} from "@/_services";
 
     import {defineComponent} from "vue";
-    import {useCommunityStore} from "@/pinia/communityStore";
+    import {useCommunityStore} from "@/store/communityStore";
 
     export default defineComponent({
         name: "ViewGrandPrixList",
