@@ -32,8 +32,8 @@
     import UserLevelResume from "@/components/user/UserLevelResume.vue";
 
     import {defineComponent} from "vue";
-    import {useAuthStore} from "@/pinia/authStore";
-    import {useCommunityStore} from "@/pinia/communityStore";
+    import {useAuthStore} from "@/store/authStore";
+    import {useCommunityStore} from "@/store/communityStore";
 
     export default defineComponent({
         name: "Home",
