@@ -47,8 +47,8 @@
     import CommunitiesDropdown from "@/components/navbar/CommunitiesDropdown.vue";
 
     import {defineComponent} from "vue";
-    import {useAuthStore} from "@/pinia/authStore";
-    import {useCommunityStore} from "@/pinia/communityStore";
+    import {useAuthStore} from "@/store/authStore";
+    import {useCommunityStore} from "@/store/communityStore";
 
     export default defineComponent({
         name: "LandingNavbar",
