@@ -121,8 +121,8 @@ import {competitionService} from "@/_services";
 import {RuleSet} from "@/types/RuleSet";
 import {isValidCommunity} from "@/utils";
 import {defineComponent} from "vue";
-import {useAuthStore} from "@/pinia/authStore";
-import {useCommunityStore} from "@/pinia/communityStore";
+import {useAuthStore} from "@/store/authStore";
+import {useCommunityStore} from "@/store/communityStore";
 
 export default defineComponent({
     name: "RulesComponent",
