@@ -30,7 +30,7 @@
 </template>
 
 <script lang="ts">
-import {useAuthStore} from "@/pinia/authStore";
+import {useAuthStore} from "@/store/authStore";
 import {defineComponent} from "vue";
 
 export default defineComponent({
