@@ -1,4 +1,4 @@
-import Oruga from "@oruga-ui/oruga";
+import Oruga from "@oruga-ui/oruga-next";
 
 declare module 'vue/types/vue' {
     interface Vue {
