@@ -1,9 +1,9 @@
 import { defineConfig } from 'vite'
-import { createVuePlugin } from 'vite-plugin-vue2'
 import {fileURLToPath} from "url";
+import vue from '@vitejs/plugin-vue'
 
 export default defineConfig({
-    plugins: [createVuePlugin()],
+    plugins: [vue()],
     resolve: {
         alias: {
             '@': fileURLToPath(new URL('./src', import.meta.url))
